==> 00_P3_CS.Rmd <==
---
title: "CS_P3"
author: "Boyana Buyuklieva"
date: "September 26, 2022"
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache.lazy = FALSE)
# library(dplyr)
# library(tidyverse)
# library(ggExtra)
# library(gridExtra)

library(data.table)
library(readr)
# loads: loadRData
source("../Outputs/Thesis_custom_scripts.R", local = knitr::knit_global())

#All of these contain entries where the UPRN is NULL, but the address is unique
#sum(duplicated(big_SC_dt$n_UPRN)) == sum(is.na(big_SC_dt$n_UPRN)) + 1
big_SC_dt<- loadRData( "../Data/RData_crossSectional/big_crossSect_EPCs_cleaned_N16m984358.RData") 
additional_dt <- loadRData( "../Data/RData_crossSectional/big_crossSect_EPCs_cleaned_N16m984358_additional_dt.RData") 


combi_SC_dt <-merge(big_SC_dt, additional_dt, by = 'LMK_KEY', all = TRUE)
```


# Edits to EE & EI
```{r EE & EI checks}
#formatting:
lettermapping <-  data.frame( letter = c('A','B','C','D','E','F','G'), 
                              cutoff = c(92, 81, 69, 55, 39, 21, 0))

#convert to numeric
combi_SC_dt[,CURRENT_ENERGY_EFFICIENCY := as.integer(CURRENT_ENERGY_EFFICIENCY)]
combi_SC_dt[,POTENTIAL_ENERGY_EFFICIENCY:=as.integer(POTENTIAL_ENERGY_EFFICIENCY)]

combi_SC_dt[,ENVIRONMENT_IMPACT_CURRENT:=as.integer(ENVIRONMENT_IMPACT_CURRENT)]
combi_SC_dt[,ENVIRONMENT_IMPACT_POTENTIAL:=as.integer(ENVIRONMENT_IMPACT_POTENTIAL)]

#No void entries
sum(is.na(combi_SC_dt$CURRENT_ENERGY_EFFICIENCY))
sum(is.na(combi_SC_dt$POTENTIAL_ENERGY_EFFICIENCY))
sum(is.na(combi_SC_dt$ENVIRONMENT_IMPACT_CURRENT))
sum(is.na(combi_SC_dt$ENVIRONMENT_IMPACT_POTENTIAL))

# Invalid entries [n = 73] are some with EE = 0, and more with EE > 100.
combi_SC_dt[ , by = CURRENT_ENERGY_RATING, .N ]
combi_SC_dt[ CURRENT_ENERGY_EFFICIENCY == 0, .N] # 7 zero entries. 
combi_SC_dt[ CURRENT_ENERGY_EFFICIENCY >  100 , .N] 
#View(combi_SC_dt[ CURRENT_ENERGY_RATING == 'INVALID!', list(CURRENT_ENERGY_EFFICIENCY,ENVIRONMENT_IMPACT_CURRENT) ])

# Invalid entries are some with EI = 0
# combi_SC_dt[ ENVIRONMENT_IMPACT_CURRENT == '0', .N ] # 7 
```


```{r EE & EI Adjusting values over 100, the A+}
# round(quantile(combi_SC_dt$CURRENT_ENERGY_EFFICIENCY, 0.99, na.rm = T),0) #87 
# round(quantile(combi_SC_dt$POTENTIAL_ENERGY_EFFICIENCY, 0.99, na.rm = T),0) #97 
# round(quantile(combi_SC_dt$ENVIRONMENT_IMPACT_CURRENT, 0.99, na.rm = T),0) #93 
# round(quantile(combi_SC_dt$ENVIRONMENT_IMPACT_POTENTIAL, 0.99, na.rm = T),0) #100 

#Edit 1: CEE values over 100 are set to 101
combi_SC_dt[, n_CURRENT_ENERGY_EFFICIENCY := ifelse( CURRENT_ENERGY_EFFICIENCY>100 ,101, CURRENT_ENERGY_EFFICIENCY)]
combi_SC_dt[, n_POTENTIAL_ENERGY_EFFICIENCY := ifelse( POTENTIAL_ENERGY_EFFICIENCY>100 ,101, POTENTIAL_ENERGY_EFFICIENCY)]

#Edit 2: EIC values over 100 are set to 101
combi_SC_dt[, n_ENVIRONMENT_IMPACT_CURRENT := ifelse( ENVIRONMENT_IMPACT_CURRENT>100 ,101, ENVIRONMENT_IMPACT_CURRENT)]
combi_SC_dt[, n_ENVIRONMENT_IMPACT_POTENTIAL := ifelse( ENVIRONMENT_IMPACT_POTENTIAL>100 ,101, ENVIRONMENT_IMPACT_POTENTIAL)]


#Edited Variable 1: Potential Change PEE 
combi_SC_dt[, nn_potEE := n_POTENTIAL_ENERGY_EFFICIENCY - n_CURRENT_ENERGY_EFFICIENCY]
#Edited Variable 2: Potential Change EIP 
combi_SC_dt[, nn_potIE := n_ENVIRONMENT_IMPACT_POTENTIAL - n_ENVIRONMENT_IMPACT_CURRENT]
```


```{r EE & EI create_Ratings}
#Edit 8: Difference in CEE & CEI
#note: the CURRENT_ENERGY_RATING  includes A+ and 0 ratings, that are marked as 'INVALID!' 
combi_SC_dt[ , nn_CEE_rating := cut(CURRENT_ENERGY_EFFICIENCY,
                            breaks = c(rev(lettermapping$cutoff), Inf),
                            right = F, #includes the lower value
                            labels=  rev(lettermapping$letter)) ]

combi_SC_dt[ , nn_CEI_rating := cut(ENVIRONMENT_IMPACT_CURRENT,
                            breaks = c(rev(lettermapping$cutoff), Inf),
                            right = F, #includes the lower value
                            labels=  rev(lettermapping$letter)) ]
```


```{r EE & EI C-relative}
#N2: create CEE relative column
fctr_levels <- c("C+","C","C-")

#Grouped CEE
combi_SC_dt[ , n_CEErel:= ifelse(CURRENT_ENERGY_RATING == 'C', 'C', 
                                       ifelse(CURRENT_ENERGY_RATING %in% c('A', 'B'), 'C+', 
                                              ifelse(CURRENT_ENERGY_RATING %in% c('D', 'E','F','G'),'C-','NA' )))]
combi_SC_dt[n_CEErel =="NA", n_CEErel:= ifelse(CURRENT_ENERGY_EFFICIENCY > 100, 'C+', 'C-')]
combi_SC_dt[ ,n_CEErel := factor(n_CEErel, levels = fctr_levels, ordered = T)]


#Grouped CEI
combi_SC_dt[, nn_CEIrel:= ifelse(nn_CEI_rating == 'C', 'C', 
                                       ifelse(nn_CEI_rating %in% c('A', 'B'), 'C+', 
                                              ifelse(nn_CEI_rating %in% c('D', 'E','F','G'),'C-','NA' )))]
combi_SC_dt[nn_CEIrel =="NA", nn_CEIrel:= ifelse( nn_CEI_rating > 100, 'C+', 'C-')]
combi_SC_dt[,n_CEIrel := factor(n_CEErel, levels = fctr_levels, ordered = T)]
```

```{r EE & EI CEE_CEI_variable}

combi_SC_dt[ , nn_letter_Diff_CEE_EIC := ifelse(nn_CEE_rating == nn_CEI_rating, 'same letter rate', 
                                                         ifelse(CURRENT_ENERGY_EFFICIENCY > ENVIRONMENT_IMPACT_CURRENT, 'Better Running', 'Less C02') ) ]
```



# Edits to Rooms
```{r rooms}
##Formatting Rooms:
combi_SC_dt[, NUMBER_HABITABLE_ROOMS:=as.integer(NUMBER_HABITABLE_ROOMS)]
combi_SC_dt[, NUMBER_HEATED_ROOMS:=as.integer(NUMBER_HEATED_ROOMS)]

# #View(combi_SC_dt[, .(n = .N) ,by=NUMBER_HABITABLE_ROOMS ])
# #View(combi_SC_dt[, .(n = .N) ,by=NUMBER_HEATED_ROOMS ])
# round(quantile(combi_SC_dt$NUMBER_HABITABLE_ROOMS, 0.99, na.rm = T),0) # 10
# round(quantile(combi_SC_dt$NUMBER_HEATED_ROOMS, 0.99, na.rm = T),0) # 9


#Edit 3: Places with over 7 habitable rooms are grouped together into 7+
fctr_levels <- c(paste0(seq(1,9,1), 'HaR'), '+10HaR' )
combi_SC_dt[ , n_habitable_rooms_grp := ifelse( NUMBER_HABITABLE_ROOMS>9,'+10HaR', paste0(NUMBER_HABITABLE_ROOMS,'HaR'))]
combi_SC_dt[ , n_habitable_rooms_grp := factor(n_habitable_rooms_grp, levels = fctr_levels, ordered = T)]

#Edit 4: Places with over 9 heated rooms are grouped together into 8+
fctr_levels <- c(paste0(seq(1,8,1), 'HtR'), '+9HtR' )
combi_SC_dt[ , n_heated_rooms_grp := ifelse( NUMBER_HEATED_ROOMS>8,'+9HtR', paste0(NUMBER_HEATED_ROOMS,'HtR'))]
combi_SC_dt[ , n_heated_rooms_grp := factor(n_heated_rooms_grp, levels = fctr_levels, ordered = T)]


```

# Edits to Area
```{r floorarea}
combi_SC_dt[, TOTAL_FLOOR_AREA:=as.integer(TOTAL_FLOOR_AREA)]

round(quantile(combi_SC_dt$TOTAL_FLOOR_AREA, 0.99, na.rm = T),0) #max:276m2
round(quantile(combi_SC_dt$TOTAL_FLOOR_AREA, 0.01, na.rm = T),0) #min:25m2

#Edit 5: Floor Area. 
#Note min bedroom size is 6.5m2 = 7m2
#Max is 276, based on the 99 percentile of the entries
maxthreshold <- round(quantile(combi_SC_dt$TOTAL_FLOOR_AREA, 0.99, na.rm = T),0) 
print(paste('The 99 percentile of floor area is: ', maxthreshold, 'm2'))
combi_SC_dt[, n_TOTAL_FLOOR_AREA  := ifelse( TOTAL_FLOOR_AREA < 6 ,6, TOTAL_FLOOR_AREA)]
combi_SC_dt[, n_TOTAL_FLOOR_AREA  := ifelse( n_TOTAL_FLOOR_AREA > maxthreshold , maxthreshold+1, n_TOTAL_FLOOR_AREA)]

```


Tenure, Built form, Age & Transaction
#Editing Variables should be done in the excel sheets!
```{r change_col_based_on_LUTs}
#Editing Variables should be done in the excel sheets!
#get loopUp
LUT_TENURE <- read_excel("../Data/_metaData/OS-N_EPC_v10_LUT+dictionaries.xlsx", sheet = "TENURE",range = cell_cols("B:C"))
LUT_PROPERTY_TYPE <- read_excel("../Data/_metaData/OS-N_EPC_v10_LUT+dictionaries.xlsx", sheet = "PROPERTY_TYPE",range = cell_cols("B:C"))


#TENURE
LUT_colum <- LUT_TENURE[match(combi_SC_dt$TENURE, LUT_TENURE$TENUREfactors),"harmonisation"]
combi_SC_dt[,n_tenure := LUT_colum]
combi_SC_dt[is.na(n_tenure),n_tenure := 'unknown'] #deals with NAs
fct <- c('owner-occupied','rental (private)','rental (social)','unknown')
combi_SC_dt[,n_tenure := factor(n_tenure, levels = fct, ordered = T)]

#PROPERTY_TYPE
LUT_colum <- LUT_PROPERTY_TYPE[match(combi_SC_dt$PROPERTY_TYPE, LUT_PROPERTY_TYPE$PROPERTY_TYPEfactors),"harmonisation"]
combi_SC_dt[,n_propTyp := LUT_colum]
combi_SC_dt[is.na(n_propTyp),n_propTyp := 'unknown'] 
fct <- c('Flat','House','Bungalow','Park home')
combi_SC_dt[,n_propTyp := factor(n_propTyp, levels = fct, ordered = T)]
```

```{r LUT_TRANSACTION_TYPE}
LUT_TRANSACTION_TYPE <- read_excel("../Data/_metaData/OS-N_EPC_v10_LUT+dictionaries.xlsx", sheet = "TRANSACTION_TYPE",range = cell_cols("B:E"))

#TRANSACTION_TYPE
LUT_colum <- LUT_TRANSACTION_TYPE[match(combi_SC_dt$TRANSACTION_TYPE, LUT_TRANSACTION_TYPE$TRANSACTION_TYPEfactors),"harmonisation"]
combi_SC_dt[,n_transactTyp := LUT_colum]
combi_SC_dt[is.na(n_transactTyp),n_transactTyp := 'unknown'] 

#TRANSACTION_TYPE
LUT_colum <- LUT_TRANSACTION_TYPE[match(combi_SC_dt$n_transactTyp, LUT_TRANSACTION_TYPE$TRANSACTION_TYPEharm),"TRANSACTION_TYPE_harmSimple"]
combi_SC_dt[,nn_transactTyp := LUT_colum]
combi_SC_dt[is.na(nn_transactTyp),nn_transactTyp := 'unknown']  
```


```{r builtform}
#N: create new Built Form column
#Old levels: "unknown","End-Terrace","Detached","Mid-Terrace", "Semi-Detached"
LUT_BUILT_FORM <- read_excel("../Data/_metaData/OS-N_EPC_v10_LUT+dictionaries.xlsx", sheet = "BUILT_FORM",range = cell_cols("B:D"))
LUT_colum <- LUT_BUILT_FORM[match(combi_SC_dt$BUILT_FORM, LUT_BUILT_FORM$BUILT_FORMfactors),"harmonisation_det"]
combi_SC_dt[,n_builtForm := LUT_colum]
combi_SC_dt[is.na(n_builtForm),n_builtForm := 'unknown'] 
fct <- c('Detached','Semi-Detached','End-Terrace','Mid-Terrace','unknown')
combi_SC_dt[,n_builtForm := factor(n_builtForm, levels = fct, ordered = T)]


LUT_colum <- LUT_BUILT_FORM[match(combi_SC_dt$BUILT_FORM, LUT_BUILT_FORM$BUILT_FORMfactors),"harmonisation"]
combi_SC_dt[,nn_builtForm := LUT_colum]
combi_SC_dt[is.na(nn_builtForm),nn_builtForm := 'unknown'] 
fct <- c('Detached','Semi-Detached','Terrace','unknown')
combi_SC_dt[,n_builtForm := factor(nn_builtForm, levels = fct, ordered = T)]

#View( combi_SC_dt[,.(n= .N), by = n_builtForm])
#View( combi_SC_dt[,.(n= .N), by = nn_builtForm])
```



```{r age band}
#AGE_BAND
#age from 2007 is ambigious
LUT_AGE_BAND <- read_excel("../Data/_metaData/OS-N_EPC_v10_LUT+dictionaries.xlsx", sheet = "CONSTRUCTION_AGE_BAND",range = cell_cols("B:D"))
LUT_colum <- LUT_AGE_BAND[match(combi_SC_dt$CONSTRUCTION_AGE_BAND, LUT_AGE_BAND$CONSTRUCTION_AGE_BANDfactors), "harmonisation_det"]
combi_SC_dt[,n_construction_age_band := LUT_colum]
combi_SC_dt[is.na(n_construction_age_band),n_construction_age_band := 'unknown'] #deals with NAs
fct <- c("Pre-1900","1900-1929","1930-1949","1950-1966","1967-1975","1976-1982","1983-1990","1991-1995","1996-2002","2003-2006",
         "2007-2011","unknown","2007 onwards","2012 onwards")
combi_SC_dt[,n_construction_age_band := factor(n_construction_age_band, levels = fct, ordered = T)]


LUT_colum <- LUT_AGE_BAND[match(combi_SC_dt$CONSTRUCTION_AGE_BAND, LUT_AGE_BAND$CONSTRUCTION_AGE_BANDfactors),"harmonisation"]
combi_SC_dt[,nn_construction_age_band := LUT_colum]
combi_SC_dt[is.na(nn_construction_age_band),nn_construction_age_band := 'unknown'] #deals with NAs
fct <- c("Pre-1900","1900-1929","1930-1949","1950-1966","1967-1975","1976-1982","1983-1990","1991-1995","1996-2002","2003-2006","2007 onwards","unknown")
combi_SC_dt[,nn_construction_age_band := factor(nn_construction_age_band, levels = fct, ordered = T)]

#View(combi_SC_dt[, .(n = .N) ,by=n_construction_age_band ])
#View(combi_SC_dt[, .(n = .N) ,by=nn_construction_age_band ])
```




```{r attach_ONS_geo}
rm(additional_dt)
#Load and shrink ONS Geography
ONSUD_dt <- loadRData( "../Data/RData_crossSectional/big_crossSect_EPCs_cleaned_additional_ONSUD_dt16M293951.RData") 
ONSUD_cols <- read_csv("../Data/RData_crossSectional/ONSUD_cols.csv")
ONSUD_cols <- ONSUD_cols$cols[!is.na(ONSUD_cols$b_use)] 
ONSUD_dt <- ONSUD_dt[,ONSUD_cols , with = F]

combi_SC_dt <-merge(combi_SC_dt, ONSUD_dt, by = 'LMK_KEY', all = TRUE)
rm(ONSUD_dt)
```


```{r attach_IMDs-England}
#income deprivation E
income_imd2019lsoa_E <- fread("../Data/imd2019lsoa/imd2019lsoa.csv",
                     drop="Units", #empty column anyway
                     colClasses = c( "FeatureCode" = "character",
                                     "Measurement" = "character",
                                     "Value" = "numeric",
                                     "Indices of Deprivation" = "character"))
#Measurements are: "Rank"   "Decile" "Score"
#income_imd2019lsoa_E <- setNames(income_imd2019lsoa_E, c("lsoa11cd","Date","Measurement","Value","Indices of Deprivation") )
setnames(income_imd2019lsoa_E, c("FeatureCode"), c("lsoa11cd"))
#income_imd2019lsoa_E[ , by = `Indices of Deprivation`, .N ]



#Add IoD
##can select other IoD here
#b.income
imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "b. Income Deprivation Domain",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_bIoDRank_income','n_bIoDDecil_income') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_income_grp := cut(n_bIoDDecil_income,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('IncomeDeprived_1-3','Average.4-7','Affluent_8-10'))]


#g. Barriers to Housing and Services
imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "g. Barriers to Housing and Services Domain",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_gIoDRank_barriers','n_gIoDDecil_barriers') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_barriers_grp := cut(n_gIoDDecil_barriers,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('H&S_Deprived_1-3','Average.4-7','Affluent_8-10'))]


#i. Income Deprivation Affecting Children Index (IDACI)
imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "i. Income Deprivation Affecting Children Index (IDACI)",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_iIoDRank_IDACI','n_iIoDDecil_IDACI') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_IDACI_grp := cut(n_iIoDDecil_IDACI,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived_Child_1-3','Average.4-7','Affluent_Child_8-10'))]



#j. Income Deprivation Affecting Older People Index (IDAOPI)
imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "j. Income Deprivation Affecting Older People Index (IDAOPI)",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_jIoDRank_IDAOPI','n_jIoDDecil_IDAOPI') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_IDAOPI_grp := cut(n_jIoDDecil_IDAOPI,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived_Older_1-3','Average.4-7','Affluent_Older_8-10'))]



#a. Index of Multiple Deprivation (IMD)
imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "a. Index of Multiple Deprivation (IMD)",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_imdRank','n_imdDecil') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_imd_grp := cut(n_imdDecil,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived.1-3','Average.4-7','M.Affluent.8-10'))]






#c. Employment Deprivation Domain
imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "c. Employment Deprivation Domain",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_cIoDRank_employ','n_cIoDDecil_employ') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_employ_grp := cut(n_cIoDDecil_employ,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived.1-3_employ','Average.4-7','Affluent_8-10'))]


#d. Education, Skills and Training Domain
imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "d. Education, Skills and Training Domain",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_dIoDRank_ed','n_dIoDDecil_edu') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_edu_grp := cut(n_dIoDDecil_edu,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived.1-3_edu','Average.4-7','Affluent_8-10'))]



imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "h. Living Environment Deprivation Domain",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_hIoDRank_livEnv','n_hIoDDecil_livEnv') := .(Rank,Decile)]

#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_livEnv_grp := cut(n_hIoDDecil_livEnv,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived.1-3_livEnv','Average.4-7','Affluent_8-10'))]


imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "e. Health Deprivation and Disability Domain",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_eIoDRank_health','n_eIoDDecil_health') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_health_grp := cut(n_eIoDDecil_health,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived.1-3_health','Average.4-7','Affluent_8-10'))]


imd2019col <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "f. Crime Domain",]
imd2019col <-dcast(imd2019col, lsoa11cd ~ Measurement, value.var = c("Value"))
combi_SC_dt[imd2019col,  on = 'lsoa11cd', c('n_fIoDRank_crime','n_fIoDDecil_crime') := .(Rank,Decile)]
#Make barriers deprivation groups
combi_SC_dt[ , n_IoDDecil_crime_grp := cut(n_fIoDDecil_crime,
                            breaks = c(0, 3, 7, Inf),
                            include.lowest= TRUE, labels=c('Deprived.1-3_crime','Average.4-7','Affluent_8-10'))]

```



```{r interactions_and_simplications}

combi_SC_dt[ ,x_nn_builtFormXn_propTyp:= paste(nn_builtForm,n_propTyp)]
combi_SC_dt[ ,x_n_tenureXnn_builtFormXn_propTyp:= paste(n_tenure,x_nn_builtFormXn_propTyp)]

#simplified
#View(combi_SC_dt[, .(n = .N) ,by=x_nn_builtFormXn_propTyp ]) # most common is Terrace House
combi_SC_dt[ ,x_simp_nn_builtFormXn_propTyp:= x_nn_builtFormXn_propTyp ]
combi_SC_dt[ ,x_simp_nn_builtFormXn_propTyp:= ifelse( grepl("Flat", x_nn_builtFormXn_propTyp), 'Flat', x_simp_nn_builtFormXn_propTyp)]
combi_SC_dt[ ,x_simp_nn_builtFormXn_propTyp:= ifelse( grepl("Bungalow", x_nn_builtFormXn_propTyp), 'Bungalow', x_simp_nn_builtFormXn_propTyp)]
combi_SC_dt[ ,x_simp_nn_builtFormXn_propTyp:= ifelse( grepl("Park home", x_nn_builtFormXn_propTyp), 'Park home', x_simp_nn_builtFormXn_propTyp)]
fct <- c('Flat','Terrace House','Semi-Detached House','Detached House',"Bungalow","unknown House","Park home")
combi_SC_dt[,x_simp_nn_builtFormXn_propTyp := factor(x_simp_nn_builtFormXn_propTyp, levels = fct, ordered = T)]
#View(combi_SC_dt[, .(n = .N) ,by=x_simp_nn_builtFormXn_propTyp ])



#NOTE faster: DT[, setattr(Species, "levels", c("set", "ver", "vir"))]
combi_SC_dt[,simp_nn_construction_age_band := nn_construction_age_band]
fct <- c("Pre-1900","Early Interwar","Near WW2","Post WW2","Mid60s-90s","Mid60s-90s","Mid60s-90s","Post 1990s","Pos=t 1990s","Post 1990s","Post GFC","unknown")
levels(combi_SC_dt$simp_nn_construction_age_band) <- fct
combi_SC_dt[,simp_nn_construction_age_band := factor(simp_nn_construction_age_band, levels = unique(fct), ordered = T)]
#View(combi_SC_dt[, .(n = .N) ,by=c('nn_construction_age_band','simp_nn_construction_age_band') ])



combi_SC_dt[,ssimp_nn_construction_age_band := nn_construction_age_band]
fct <- c("Victorian and earlier (pre 1900s)",
         "Edwardian (1900-29)",
         "Circa WW2 (1930-49)",
         "Post WW2 (1950-66)",
         "Modern (1967-82)", "Modern (1967-82)",
         "Post Modern (1983-95)","Post Modern (1983-95)",
         "New (1996-06)","New (1996-06)",
         "From 2007 onwards","From 2007 onwards", 
         "From 2007 onwards","Not applicable")
levels(combi_SC_dt$ssimp_nn_construction_age_band) <- fct
combi_SC_dt[,ssimp_nn_construction_age_band := factor(ssimp_nn_construction_age_band, levels = unique(fct), ordered = T)]
#View(combi_SC_dt[, .(n = .N) ,by=c('nn_construction_age_band','simp_nn_construction_age_band', 'ssimp_nn_construction_age_band') ])
```



```{r ruc11ind_grp}
#NOTE!
combi_SC_dt[ ,ruc11ind_grp:= 'rural']
combi_SC_dt[ ,ruc11ind_grp:= ifelse( ruc11ind == 'A1', 'urban', ruc11ind_grp)]
combi_SC_dt[ ,ruc11ind_grp:= ifelse( ruc11ind == 'B1', 'urban', ruc11ind_grp )]
combi_SC_dt[ ,ruc11ind_grp:= ifelse( ruc11ind == 'C1', 'urban', ruc11ind_grp )]
combi_SC_dt[ ,ruc11ind_grp:= ifelse( ruc11ind == 'C2', 'urban', ruc11ind_grp )]

#View(head(combi_SC_dt[ ,c('ruc11ind_grp','ruc11ind'), with = F], 20))
```


```{r}
if(F){
#Date added: Nov 10th
combi_SC_dt[ ,x_PropT3_Tenure3_Age8:= paste(ssimp_nn_construction_age_band,"x",n_propTyp,"x",n_tenure)]
combi_SC_dt[ ,x_PropT3_Age8:= paste(n_propTyp,"x",ssimp_nn_construction_age_band)]

combi_SC_dt[ ,x_PropT3_Tenure3:= paste(n_propTyp,"x",n_tenure)]
combi_SC_dt[ ,X_Tenure3_Age8:= paste(n_tenure,"x",ssimp_nn_construction_age_band)]

combi_SC_dt[ ,x_Tenure3_habitable_rooms_grp10:= paste(n_tenure,"x", n_habitable_rooms_grp)]
combi_SC_dt[ ,x_Tenure3_habitable_rooms_grp10_PropT3:= paste(n_tenure,"x", n_propTyp,"x", n_habitable_rooms_grp)]
}



combi_SC_dt[,n_CEErel_binary := n_CEErel]
fct <- c("C & C+","C & C+","C-")
levels(combi_SC_dt$n_CEErel_binary) <- fct
combi_SC_dt[,n_CEErel_binary := factor(n_CEErel_binary, levels = unique(fct), ordered = T)]

combi_SC_dt[ ,x_Tenure3_CEE_binary:= paste(n_tenure,"x", n_CEErel_binary)]



combi_SC_dt[ , EXTENSION_COUNT:= as.numeric(EXTENSION_COUNT), ]
```


```{r SAP_climatic_Regions}
#SAP Climatic Regions - move to 00_P3
climatic <- as.data.table(read_csv("../Data/m_SAP_climatic_May_2020_combined.csv"))
climatic$lad19cd <- climatic$lad20cd 
combi_SC_dt[climatic,  on = 'lad19cd', SAP_climate_grp := SAP_climat]

#Add SAP version
combi_SC_dt[ , SAP_version := year(INSPECTION_DATE)]
#contains some INSPECTION_DATE nas
combi_SC_dt[ , SAP_version := ifelse(SAP_version < 2005 , 'SAP05', ifelse(SAP_version < 2009, 'SAP09', 'SAP12'))]
```

```{r Adjust Age}
#Adjust Age
#possibly remove nn_construction_age_band
combi_SC_dt[ , inspection_yr := year(INSPECTION_DATE)]
#1 issue: '216525930262009012815361707988770', 2007 onward, but inspected in 2000
#View(combi_SC_dt[ ,  .(n = .N), by = c('n_construction_age_band','inspection_yr')])
combi_SC_dt[ , nnn_construction_age_band12 := n_construction_age_band]
combi_SC_dt[ nnn_construction_age_band12 == "2007 onwards", nnn_construction_age_band12 := ifelse(inspection_yr < 2012, '2007-2011', ifelse( inspection_yr > 2011, "2012 onwards" ,"unknown" ) )]
combi_SC_dt[ , nnn_construction_age_band12 := droplevels(nnn_construction_age_band12)] #rm,"2007 onwards"
#might need to relevel
```

```{r}
#Resolves issue of LAD and RGNs, some 4k rows are still unresolved. See TODO folder

#Assign the most recent LA, based on UPRN data
combi_SC_dt[,m_EPC_lad19cd := ifelse(is.na(lad19cd), LOCAL_AUTHORITY, lad19cd)]
combi_SC_dt[,m_EPC_lad19cd := ifelse(m_EPC_lad19cd == "", NA, m_EPC_lad19cd)]
#sum(is.na(combi_SC_dt$m_EPC_lad19cd)) #check that = 0
#remove "LOCAL_AUTHORITY" from the folders, but save a boolean
combi_SC_dt[,tmp_EPC_LA_folder_info := ifelse( lad19cd == LOCAL_AUTHORITY, 'lad19cd == LOCAL_AUTHORITY', LOCAL_AUTHORITY)]
combi_SC_dt[,LOCAL_AUTHORITY := NULL]
combi_SC_dt[,LAD11NM := NULL]#realted to LOCAL_Authority
combi_SC_dt[,RGN11NM := NULL]

#Add the new geogs: LAD19, RGN19, CTRY19
lad_rgn_2019 <- as.data.table(read_csv("../Data/_shapefiles/lad19cd-rgn19nm/m_lad_rgn_2019.csv"))
setkey(lad_rgn_2019, m_EPC_lad19cd)
setkey(combi_SC_dt, m_EPC_lad19cd)
combi_SC_dt <- lad_rgn_2019[combi_SC_dt, on = 'm_EPC_lad19cd',]
```
```{r _LUT_GLAZING}
#Editing Variables should be done in the excel sheets!
LUT_GLAZING <- read_excel("../Data/_metaData/OS-N_EPC_v10_LUT+dictionaries.xlsx", sheet = "GLAZED_TYPE",range = cell_cols("B:C"))

LUT_colum <- LUT_GLAZING[match(tolower(NS_dt$GLAZED_TYPE), LUT_GLAZING$GLAZED_TYPEfactors),"harmonisation"]
NS_dt[,n_glazing := LUT_colum]
NS_dt[is.na(n_glazing),n_glazing := 'unknown'] #deals with NAs

#relevel
fct <- c('unknown','single glazing','triple glazing','double glazing','secondary glazing')
NS_dt[ ,n_glazing := factor(n_glazing, levels = fct, ordered = T)]
#View(head(NS_dt[,c('n_glazing','GLAZED_TYPE')], 300 ))
```

```{r nn_transactTyp}
fct <- c("sale","rental","srental","new","assessment","survey")
NS_dt[ ,nn_transactTyp  := factor(nn_transactTyp , levels = fct, ordered = T)]

fct <- c("marketed sale", "non marketed sale", "rental (private)", "rental (social)", "rental","new dwelling", "ECO assessment", "assessment for green deal", "FiT application", "RHI application","following green deal", "stock condition survey", "not sale or rental", "unknown")
NS_dt[ ,n_transactTyp  := factor(n_transactTyp , levels = fct, ordered = T)]
```

```{r LUT_FUEL}
#Editing Variables should be done in the excel sheets!
LUT_MFUEL <- read_excel("../Data/_metaData/OS-N_EPC_v10_LUT+dictionaries.xlsx", sheet = "MAIN_FUEL",range = cell_cols("B:D"))


#MAIN FUEL
LUT_colum <- LUT_MFUEL[match(tolower(NS_dt$MAIN_FUEL), LUT_MFUEL$MAIN_FUELfactors),"harmonisation"]
NS_dt[,n_mFuel := LUT_colum]
NS_dt[is.na(n_mFuel), n_mFuel := 'unspecified'] #deals with NAs
NS_dt[ ,n_mFuel:= relevel(as.factor(n_mFuel), 'mains gas')]

LUT_colum <- LUT_MFUEL[match(tolower(NS_dt$MAIN_FUEL), LUT_MFUEL$MAIN_FUELfactors),"m_mFuel_community_bool"]
NS_dt[,n_mFuel_community_bool := LUT_colum]
NS_dt[is.na(n_mFuel_community_bool), n_mFuel_community_bool := NA ] #deals with NAs

#Simplified Fuel 
NS_dt[,nn_mFuel9 := n_mFuel]
NS_dt[nn_mFuel9 == 'fBlend' , nn_mFuel9 := 'oil']


#View(head(NS_dt[,c('n_mFuel','nn_mFuel9','MAIN_FUEL','n_mFuel_community_bool')], 300 ))
```

```{r deprivation_ranks_to_percentiles}
#"n_imdRank"  -  > "n_imdPercent"
#‘percentage’ measure rather than the ‘rank’ measure 
#'IoDPercent_'
rank_names <- names(NS_dt)[grep('Rank',names(NS_dt))]
percent_names <- gsub('Rank', 'Percent', rank_names)
NS_dt <- NS_dt[ , (percent_names):= lapply(.SD, function(x) rank(x)/length(x)), .SDcols = rank_names]
```

```{r agebands}
#set the newest homes to be the baseline
NS_dt <- droplevels(NS_dt)#drop unused levels, takes ages; needed for age bands below
NS_dt[ ,nnn_construction_age_band12:= forcats::fct_rev(nnn_construction_age_band12) ]


NS_dt[ , nnn_construction_age_band10 := nnn_construction_age_band12]
NS_dt[ nnn_construction_age_band10 == "1983-1990"| nnn_construction_age_band10 == "1991-1995", nnn_construction_age_band10 := "1983-1995"]
NS_dt[ nnn_construction_age_band10 == "1950-1966" | nnn_construction_age_band10 == "1967-1975", nnn_construction_age_band10 := "1950-1975"]
NS_dt[ , nnn_construction_age_band10 := droplevels(nnn_construction_age_band10)]
fct <- c("2012 onwards", "2007-2011", "2003-2006", "1996-2002",
         "1983-1995",
         "1976-1982",
         "1950-1975",
         "1930-1949","1900-1929","Pre-1900")
NS_dt[,nnn_construction_age_band10 := factor(nnn_construction_age_band10, levels = fct , ordered = T)]



NS_dt[ , nnn_construction_age_band9:= nnn_construction_age_band10]
NS_dt[ nnn_construction_age_band9 == "1983-1995", nnn_construction_age_band9 := "1976-1995"]
NS_dt[ nnn_construction_age_band9 == "1976-1982", nnn_construction_age_band9 := "1976-1995"]

fct <- c("2012 onwards", "2007-2011", "2003-2006", "1996-2002",
         "1976-1995",
         "1950-1975",
         "1930-1949","1900-1929","Pre-1900")
NS_dt[,nnn_construction_age_band9 := factor(nnn_construction_age_band9, levels = fct , ordered = T)]
```

#End
```{r}
#Note:   sum(is.na(combi_SC_dt$lsoa11cd)) # 690407 //  sum(is.na(combi_SC_dt$n_eIoDDecil_health)) # 1534228
#save(combi_SC_dt, file ="../Data/RData_crossSectional/_combi_SC_dt_big_crossSect_EPCs_cleaned_N16m984358.RData")

date <- format(Sys.Date(), '%Y_%m%d')
#Adds date #
metaDataCSV( combi_SC_dt , paste("combi_SC_dt_N16m984358_", date) , dir = "../Data/_metaData/CSV_")

#combi_SC_dt <- loadRData("../Data/RData_crossSectional/_combi_SC_dt_big_crossSect_EPCs_cleaned_N16m984358.RData")
```

```{r}
out <- sapply(combi_SC_dt, function(x){  gsub(',',' ',toString(class(x))) })
out <- cbind.data.frame(name = names(combi_SC_dt), class = out)
out$unique <- sapply(combi_SC_dt, function(x){ ifelse(typeof(x) != 'character' ,NA , length(unique(x)))})

getuniquevals_under200 <- function(x){ifelse(typeof(x) == 'character', ifelse(length(unique(x)) < 200, toString(unique(x)), NA), NA)}
out$uniques_under_200 <- sapply(combi_SC_dt[], getuniquevals_under200)
out$sample1 <- t(combi_SC_dt[1,])
out$sample2 <- t(combi_SC_dt[2,])
out$sample3 <- t(combi_SC_dt[3,])



write.csv(out, file =paste0("../Data/RData_crossSectional/combi_SC_dt_cols_",date,".csv") )  
```
