==> P3_CrossSectional_BigDt.Rmd <==
---
title: "P3_CrossSectional_EDA"
author: "Boyana Buyuklieva"
date: "August 30, 2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(tidyverse)
library(data.table)
library(readr)
library(foreign)
library(ggExtra)


loadRData <- function(fileName){
    #loads an RData file, and returns it
    load(fileName)
    get(ls()[ls() != "fileName"])
}

big_Dt_CSsubset_N6M7 <- loadRData("../Outputs/WIP_crossSectional/ONSUD_crossSectional_subset_N16744771.RData")

fctr_levels <- c("C+","C","C-")
big_Dt_CSsubset_N6M7[,n_Crel := factor(n_Crel, levels = fctr_levels, ordered = T)]


#multiple deprivation EW
IMD19<- read.dbf("../Data/_shapefiles/Lower_Super_Output_Area_(LSOA)_IMD_2019__(OSGB1936)/Lower_Super_Output_Area_(LSOA)_IMD_2019__(OSGB1936).dbf")

#income deprivation E
income_imd2019lsoa_E <- fread("../Data/imd2019lsoa/imd2019lsoa.csv",
                     drop="Units", #empty column anyway
                     colClasses = c( "FeatureCode" = "character",
                                     "Measurement" = "character",
                                     "Value" = "numeric",
                                     "Indices of Deprivation" = "character"))

#Measurements are: "Rank"   "Decile" "Score"
#income_imd2019lsoa_E <- setNames(income_imd2019lsoa_E, c("lsoa11cd","Date","Measurement","Value","Indices of Deprivation") )
setnames(income_imd2019lsoa_E, c("FeatureCode"), c("lsoa11cd"))

EDA_cols <- read_csv("../Data/EDA_colour_scheme.csv")
#EDA_cols <- subset(EDA_cols[EDA_cols$geography_name %in%  cross_data, ])

#formatting:
lettermapping <-  data.frame( letter = c('A','B','C','D','E','F','G'), 
                              cutoff = c(92, 81, 69, 55, 39, 21, 0))
```


```{r attached_IMDs}
#Add IMD
big_Dt_CSsubset_N6M7[IMD19,  on = 'lsoa11cd', c('n_IMDRank','n_IMDDecil') := .(IMDRank,IMDDecil)]

#Add IoD
##can select other IoD here
imd_bincome <-  income_imd2019lsoa_E[ `Indices of Deprivation` == "b. Income Deprivation Domain",]
imd_bincome <-dcast(imd_bincome, lsoa11cd ~ Measurement, value.var = c("Value"))
big_Dt_CSsubset_N6M7[imd_bincome,  on = 'lsoa11cd', c('n_bIoDRank_income','n_bIoDDecil_income') := .(Rank,Decile)]
```

```{r}
big_Dt_CSsubset_N6M7[, n_potCEE := POTENTIAL_ENERGY_EFFICIENCY - CURRENT_ENERGY_EFFICIENCY]
big_Dt_CSsubset_N6M7[, n_potEIC := ENVIRONMENT_IMPACT_POTENTIAL - ENVIRONMENT_IMPACT_CURRENT]


tmp <- big_Dt_CSsubset_N6M7[sample(nrow(big_Dt_CSsubset_N6M7), 2000, replace = F ), ]
```



```{r potential_energy}
start_time <- Sys.time()

counts <- sum(which( big_Dt_CSsubset_N6M7$POTENTIAL_ENERGY_EFFICIENCY>100))
percents <- round(counts/length(big_Dt_CSsubset_N6M7$POTENTIAL_ENERGY_EFFICIENCY),0)
counts <- format(counts, big.mark = ' ', trim = T)

p <- ggplot(big_Dt_CSsubset_N6M7, aes(x=POTENTIAL_ENERGY_EFFICIENCY, y = '')) +
      geom_point() +
      #geom_boxplot(width = .2)+
      geom_boxplot()+
      annotate(geom="label", 
                label= paste0("Over 100:\n",counts ,'\n ~',percents,'%'),
                x=110, y=1.2, angle = 90)

 ggMarginal(p, type="density") #,size=3)
 
 
print(Sys.time() - start_time)
```













```{r understand CEE, EIC & PEE, EIP}


ggplot(tmp, aes(x = POTENTIAL_ENERGY_EFFICIENCY)) + geom_density()+ geom_boxplot()


#https://r-graph-gallery.com/277-marginal-histogram-for-ggplot2.html


counts <- sum(which( tmp$POTENTIAL_ENERGY_EFFICIENCY>100))
percents <- round(counts/length(tmp$POTENTIAL_ENERGY_EFFICIENCY),0)
counts <- format(counts, big.mark = ' ', trim = T)

p <- ggplot(tmp, aes(x=POTENTIAL_ENERGY_EFFICIENCY, y = '')) +
      geom_point() +
      #geom_boxplot(width = .2)+
      geom_boxplot()+
      annotate(geom="label", 
                label= paste0("Over 100:\n",counts ,'\n ~',percents,'%'),
                x=110, y=1.2, angle = 90)

 ggMarginal(p, type="density") #,size=3)

```


```{r understand TFS, Heated & habitable }
counts <- sum(which( tmp$POTENTIAL_ENERGY_EFFICIENCY>100))
percents <- round(counts/length(tmp$POTENTIAL_ENERGY_EFFICIENCY),0)
counts <- format(counts, big.mark = ' ', trim = T)

p <- ggplot(tmp, aes(x=tmp$TOTAL_FLOOR_AREA, y = '')) +
      geom_point() +
      geom_boxplot()
      # +
      # annotate(geom="label", 
      #           label= paste0("Over 100:\n",counts ,'\n ~',percents,'%'),
      #           x=110, y=1.2, angle = 90)

 ggMarginal(p, type="density") #,size=3)
```
















```{r IMD IoDb comparison}
#How does IMD compare to Income ID in England Only?
##Scatter plot

#ggplot(tmp[!is.na(n_bIoDRank_income), ],
ggplot(big_Dt_CSsubset_N6M7[!is.na(n_bIoDRank_income), ],
       aes(x=n_IMDRank, y=n_bIoDRank_income, color=n_Crel)) +
  geom_point(alpha = 1/10) + 
  geom_smooth(method=lm, se=FALSE, fullrange=TRUE)+
  scale_colour_manual(values = c("C+" = "#ffa600", "C" = "#808080", "C-" = "#008cff"))


if(F){
model_Cplus <- lm(n_IMDRank ~ n_bIoDRank_income, 
            data = big_Dt_CSsubset_N6M7[!is.na(n_bIoDRank_income) & n_Crel=="C+", ])
summary(model_Cplus)

model_Cminus <- lm(n_IMDRank ~ n_bIoDRank_income, 
            data = big_Dt_CSsubset_N6M7[!is.na(n_bIoDRank_income) & n_Crel=="C-", ])
summary(model_Cminus)

model_C <- lm(n_IMDRank ~ n_bIoDRank_income, 
            data = big_Dt_CSsubset_N6M7[!is.na(n_bIoDRank_income) & n_Crel=="C", ])
summary(model_C) 
}
```



```{r CEE and EIC relationship}
#https://r-graph-gallery.com/277-marginal-histogram-for-ggplot2.html
#ggplot(tmp,
ggplot(big_Dt_CSsubset_N6M7[ ENVIRONMENT_IMPACT_CURRENT<101],
       aes(x=nn_CURRENT_ENERGY_EFFICIENCY, y=ENVIRONMENT_IMPACT_CURRENT, color=n_Crel)) +
  geom_point(alpha = 1/10) + 
  geom_smooth(method=lm, 
              se=FALSE, 
              fullrange=F)+
  scale_colour_manual(
    values = c("C+" = "#ffa600", "C" = "#808080", "C-" = "#008cff"))
```





```{r}
#https://stackoverflow.com/questions/54393927/adding-boxplot-below-density-plot

#Ridge Density plot by Decile 
#does the 1-3, 4-7, 8-10 split make sense? 

```

```{r}
#Does the difference between current and potential CEE/ EnvE differ by Decile? Ridge density. / Density

#Does the difference between current and potential EE/ EnvE differ by Decile?  Ridge density. / Density
```



```{r}
#Is a low EPCs associated with having more space in deprived areas?
#scatter plot: y = CEE / EnvE; x = ImD/IncomeID rank , colour by region 
```
